Move RecommendedChannelsList to shared components directory.
